Рефактринг + последние приготоовления.
