+++ Assets/Game/UI/ItemPickuup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb8c0b95d461b114bb725934d4e788a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
