+++ Assets/Game/UI/ItemPickuup/ItemNotificationPrefab.prefab
@@ -227,7 +227,7 @@ GameObject:
   - component: {fileID: 2292918337874513068}
   - component: {fileID: 7275471957692411433}
   m_Layer: 5
-  m_Name: NotificationPrefab
+  m_Name: ItemNotificationPrefab
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
